All tests that aren't marked as failing pass
