+++ tests/test_benchmarks.py
@@ -1,11 +1,4 @@
-""" This module runs the test suite.
-
-@todo 
-
-    Presently this runs all of the benchmarks as tests.
-    Some of the benchmarks take a longer-than-desireable amount of time to run for routine testing.
-    Perhaps each benchmark should have arguments which allow us to run quick versions in the test suite. 
-"""
+""" This module runs the benchmark test suite. """
 from .context import phaseflow
 
 
@@ -34,11 +27,6 @@ def test_adaptive_stefan_problem_benchmark():
     phaseflow.benchmarks.AdaptiveStefanProblem().run()
     
     
-def test_adaptive_convection_coupled_melting_toy_pcm_benchmark():
-
-    phaseflow.benchmarks.AdaptiveConvectionCoupledMeltingToyPCM().run()
-    
-    
 if __name__=='__main__':
     
     test_adaptive_lid_driven_cavity_benchmark()
@@ -51,4 +39,3 @@ if __name__=='__main__':
     
     test_adaptive_stefan_problem_benchmark()
     
-    test_adaptive_convection_coupled_melting_toy_pcm_benchmark()
